[fea-rs] Refactor merging of external lookups
Adds an ExternalFeatures type, and then has merging everything be a
method there.

This PR makes no functional changes, it's just moving things around
(functional changes will come afterwards)
