+++ fea-rs/src/compile/lookups.rs
@@ -35,11 +35,7 @@ use crate::{
     Kind, Opts,
 };
 
-use super::{
-    features::{AllFeatures, FeatureLookups},
-    metrics::Anchor,
-    tags,
-};
+use super::{features::AllFeatures, metrics::Anchor, tags};
 
 use contextual::{
     ContextualLookupBuilder, PosChainContextBuilder, PosContextBuilder, ReverseChainBuilder,
@@ -55,15 +51,6 @@ use gsub_builders::{
     AlternateSubBuilder, LigatureSubBuilder, MultipleSubBuilder, SingleSubBuilder,
 };
 pub(crate) use helpers::ClassDefBuilder2;
-
-// features that can be added by a feature writer
-const CURS: Tag = Tag::new(b"curs");
-const MARK: Tag = Tag::new(b"mark");
-const MKMK: Tag = Tag::new(b"mkmk");
-const ABVM: Tag = Tag::new(b"abvm");
-const BLWM: Tag = Tag::new(b"blwm");
-const KERN: Tag = Tag::new(b"kern");
-const DIST: Tag = Tag::new(b"dist");
 
 /// A simple trait for building lookups
 // This exists because we use it to implement `LookupBuilder<T>`
@@ -487,6 +474,19 @@ impl AllLookups {
         LookupId::Gpos(self.gpos.len())
     }
 
+    /// insert a sequence of lookups into the GPOS list at a specific pos.
+    ///
+    /// After calling this, any existing items after `pos` will have invalid
+    /// `LookupId`s! the caller is expected to be doing bookkeeping, and to
+    /// subsequently remap ids.
+    pub(crate) fn splice_gpos(
+        &mut self,
+        pos: usize,
+        lookups: impl IntoIterator<Item = PositionLookup>,
+    ) {
+        self.gpos.splice(pos..pos, lookups);
+    }
+
     /// Returns `true` if there is an active lookup of this kind
     pub(crate) fn has_current_kind(&self, kind: Kind) -> bool {
         self.current.as_ref().map(SomeLookup::kind) == Some(kind)
@@ -715,142 +715,6 @@ impl AllLookups {
         lookup_ids
     }
 
-    /// Returns a map that must be used to remap the ids in any features where
-    /// they were used.
-    pub(crate) fn merge_external_lookups(
-        &mut self,
-        lookups: Vec<(LookupId, PositionLookup)>,
-        features: &BTreeMap<FeatureKey, FeatureLookups>,
-        insert_markers: &HashMap<Tag, (LookupId, usize)>,
-    ) -> LookupIdMap {
-        // okay so this is a bit complicated, so a brief outline might be useful:
-        // - 'lookups' and 'features' are being passed in from the outside
-        // - `insert_markers` are optional locations at which the lookups
-        //    for a given feature should be inserted
-        // - if a feature has no marker, its lookups are appended at the end.
-        //
-        // NOTE: inserting lookups into the middle of the lookup list means that
-        // all subsequent lookup ids become invalid. As a consequence, we need
-        // to figure out the transform from old->new for any affected lookups,
-        // and remap those ids at the end.
-
-        // preflight: we have a list of the manual insert markers, but we also
-        // need to figure out where the marker-less features should go.
-        // fonttools some something fancy here, where certain features are
-        // always inserted ahead of certain other features, if they exist;
-        // otherwise features are appended.
-        //
-        // see https://github.com/googlefonts/ufo2ft/blob/16ed156bd6a8b9bc035/Lib/ufo2ft/featureWriters/baseFeatureWriter.py#L235
-        // and https://github.com/googlefonts/ufo2ft/issues/506
-
-        let mut insert_markers = insert_markers.to_owned();
-
-        // Also: in fonttools, dist/kern are added before marks, so let's do
-        // that first; and kern has to go before dist, if both is present
-        // https://github.com/googlefonts/ufo2ft/blob/16ed156bd6a8b9bc035d0/Lib/ufo2ft/featureWriters/kernFeatureWriter.py#L311
-        if let Some((kern_id, kern_pos)) = insert_markers.get(&DIST).copied() {
-            // if kern has a marker but dist doesn't, insert dist first
-            insert_markers
-                .entry(KERN)
-                .or_insert((kern_id, kern_pos - 1));
-        }
-
-        // then handle the mark features, in the same order as ufo2ft:
-        // see https://github.com/googlefonts/ufo2ft/blob/16ed156bd6/Lib/ufo2ft/featureWriters/baseFeatureWriter.py#L235
-        let order = [ABVM, BLWM, MARK, MKMK];
-        for (i, tag) in order.iter().enumerate() {
-            if let Some((id, mut pos)) = insert_markers.get(tag).copied() {
-                // if a later feature has an explicit position, put the earlier
-                // features in front of it; rev() because we're prepending each
-                // time
-                for prev_tag in order[..i].iter().rev() {
-                    if !insert_markers.contains_key(prev_tag) {
-                        pos -= 1;
-                        insert_markers.insert(*prev_tag, (id, pos));
-                    }
-                }
-            }
-        }
-
-        // finally insert dummy markers for any other feature that doesn't exist
-        // when appending, we use the last id and increment the position
-        let last_id = LookupId::Gpos(self.gpos.len());
-        // quite likely to be after EOF!
-        let mut next_pos = 1_000_000_000;
-        // for adding extras, mimic the order ufo2ft uses, which is alphabetical
-        // but grouped by feature writer. We add markers for all features,
-        // even if they dont exist in the font; we'll ignore those later
-        for feature in [CURS, KERN, DIST, ABVM, BLWM, MARK, MKMK] {
-            insert_markers.entry(feature).or_insert_with(|| {
-                next_pos += 1;
-                (last_id, next_pos)
-            });
-        }
-
-        let lookup_to_feature = features
-            .iter()
-            .filter(|(feat, _)| insert_markers.contains_key(&feat.feature))
-            .flat_map(|(feat, lookups)| lookups.iter_ids().map(|id| (id, feat.feature)))
-            .collect::<HashMap<_, _>>();
-
-        // now we want to process the lookups that had an insert marker,
-        // and we want to do it by grouping them by feature tag.
-        let mut lookups_by_feature = HashMap::new();
-        for lk in lookups {
-            let feature = lookup_to_feature.get(&lk.0).unwrap();
-            lookups_by_feature
-                .entry(*feature)
-                .or_insert(Vec::new())
-                .push(lk);
-        }
-
-        // except we want to assign the ids respecting the order of the markers
-        // in the source file, so convert to a vec and sort.
-        let mut lookups_by_feature = lookups_by_feature.into_iter().collect::<Vec<_>>();
-        lookups_by_feature.sort_by_key(|(tag, _)| insert_markers.get(tag).unwrap());
-
-        let mut map = LookupIdMap::default();
-        let mut inserted_so_far = 0;
-
-        // 'adjustments' stores the state we need to remap existing ids, if needed.
-        let mut adjustments = Vec::new();
-
-        for (tag, lookups) in lookups_by_feature {
-            let first_id = insert_markers.get(&tag).unwrap().0.to_raw();
-            // first update the ids
-            for (i, (temp_id, _)) in lookups.iter().enumerate() {
-                let final_id = LookupId::Gpos(first_id + inserted_so_far + i);
-                map.insert(*temp_id, final_id);
-            }
-            // then insert the lookups into the correct position
-            let insert_at = first_id + inserted_so_far;
-            inserted_so_far += lookups.len();
-            self.gpos
-                .splice(insert_at..insert_at, lookups.into_iter().map(|(_, lk)| lk));
-            adjustments.push((first_id, inserted_so_far));
-        }
-
-        // now based on our recorded adjustments, figure out the remapping
-        // for the existing ids. each entry in adjustment is an (index, delta)
-        // pair, where the delta applies from adjustment[n] to adjustment[n +1]
-        if !adjustments.is_empty() {
-            // add the end of the last range
-            adjustments.push((self.gpos.len(), inserted_so_far));
-        }
-        let (mut range_start, mut adjust) = (0, 0);
-        let mut adjustments = adjustments.as_slice();
-        while let Some(((next_start, next_adjust), remaining)) = adjustments.split_first() {
-            if adjust > 0 {
-                for old_id in range_start..*next_start {
-                    map.insert(LookupId::Gpos(old_id), LookupId::Gpos(old_id + adjust));
-                }
-            }
-            (range_start, adjust, adjustments) = (*next_start, *next_adjust, remaining);
-        }
-
-        map
-    }
-
     pub(crate) fn build(
         &self,
         features: &AllFeatures,
@@ -910,7 +774,7 @@ impl AllLookups {
 }
 
 impl LookupId {
-    fn to_raw(self) -> usize {
+    pub(crate) fn to_raw(self) -> usize {
         match self {
             LookupId::Gpos(idx) => idx,
             LookupId::Gsub(idx) => idx,
@@ -941,7 +805,7 @@ impl LookupId {
 }
 
 impl LookupIdMap {
-    fn insert(&mut self, from: LookupId, to: LookupId) {
+    pub(crate) fn insert(&mut self, from: LookupId, to: LookupId) {
         self.mapping.insert(from, to);
     }
 
@@ -1408,191 +1272,4 @@ fn is_gpos_rule(kind: Kind) -> bool {
             | Kind::GposType7
             | Kind::GposType8
     )
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::compile::tags::{LANG_DFLT, SCRIPT_DFLT};
-
-    use super::*;
-
-    fn make_all_lookups() -> AllLookups {
-        AllLookups {
-            gpos: (0..8)
-                .map(|_| PositionLookup::Single(Default::default()))
-                .collect(),
-            ..Default::default()
-        }
-    }
-
-    #[test]
-    fn merge_external_lookups_before() {
-        let mut all = make_all_lookups();
-
-        let lookups = (0..6)
-            .map(|id| {
-                (
-                    LookupId::External(id),
-                    PositionLookup::Pair(Default::default()),
-                )
-            })
-            .collect();
-        let features: BTreeMap<_, _> =
-            [(MARK, [0].as_slice()), (MKMK, &[1, 2]), (KERN, &[3, 4, 5])]
-                .iter()
-                .map(|(tag, ids)| {
-                    let mut features = FeatureLookups::default();
-                    features.base = ids.iter().copied().map(LookupId::External).collect();
-                    (FeatureKey::new(*tag, LANG_DFLT, SCRIPT_DFLT), features)
-                })
-                .collect();
-
-        // mark is 'after', kern is 'before', and mkmk has no marker (goes at the end)
-        let markers = HashMap::from([
-            (MARK, (LookupId::Gpos(3), 100)),
-            (KERN, (LookupId::Gpos(5), 200)),
-        ]);
-
-        let id_map = all.merge_external_lookups(lookups, &features, &markers);
-        let final_lookups = all.gpos.iter().map(|lk| lk.kind()).collect::<Vec<_>>();
-        assert_eq!(
-            final_lookups,
-            [
-                Kind::GposType1,
-                Kind::GposType1,
-                Kind::GposType1,
-                Kind::GposType2, // one mark lookup at 3
-                Kind::GposType1,
-                Kind::GposType1,
-                Kind::GposType2, // 3 kern lookups at 5
-                Kind::GposType2,
-                Kind::GposType2,
-                Kind::GposType1,
-                Kind::GposType1,
-                Kind::GposType1,
-                Kind::GposType2, // 2 mkmk lookups at the end
-                Kind::GposType2
-            ]
-        );
-        let remapped = (0..8)
-            .map(|id| id_map.get(LookupId::Gpos(id)).to_gpos_id_or_die())
-            .collect::<Vec<_>>();
-        assert_eq!(remapped, [0, 1, 2, 4, 5, 9, 10, 11]);
-        let inserted = (0..6)
-            .map(|id| id_map.get(LookupId::External(id)).to_gpos_id_or_die())
-            .collect::<Vec<_>>();
-        assert_eq!(inserted, [3, 12, 13, 6, 7, 8]);
-    }
-
-    fn mock_external_features(
-        tags: &[Tag],
-    ) -> (
-        Vec<(LookupId, PositionLookup)>,
-        BTreeMap<FeatureKey, FeatureLookups>,
-    ) {
-        let mut lookups = Vec::new();
-        let mut features = BTreeMap::new();
-
-        for (i, feature) in tags.iter().enumerate() {
-            let id = LookupId::External(i);
-            let lookup = PositionLookup::Single(Default::default());
-            let key = FeatureKey::new(*feature, LANG_DFLT, SCRIPT_DFLT);
-
-            let mut feature_lookups = FeatureLookups::default();
-            feature_lookups.base = vec![id];
-            lookups.push((id, lookup));
-            features.insert(key, feature_lookups);
-        }
-        (lookups, features)
-    }
-
-    fn make_markers_with_order<const N: usize>(order: [Tag; N]) -> HashMap<Tag, (LookupId, usize)> {
-        order
-            .into_iter()
-            .enumerate()
-            .map(|(i, tag)| (tag, (LookupId::Gpos(0), i + 10)))
-            .collect()
-    }
-
-    // after remapping, return the order of features, based on the final
-    // ordering of lookups.
-    //
-    // this works because for these tests we add one unique lookup to each feature,
-    // so lookup order is a proxy for feature order.
-    // - one lookup per feature
-    // - no duplicates
-    fn feature_order_for_test(
-        map: &LookupIdMap,
-        //ext_lookups: &[(LookupId, PositionLookup)],
-        features: &BTreeMap<FeatureKey, FeatureLookups>,
-    ) -> Vec<Tag> {
-        let mut features = features
-            .iter()
-            .map(|(key, lookups)| (key.feature, lookups.base[0]))
-            .collect::<Vec<_>>();
-        features.sort_by_key(|(_, id)| map.get(*id));
-        features.into_iter().map(|(tag, _)| tag).collect()
-    }
-
-    #[test]
-    fn feature_ordering_without_markers() {
-        let (lookups, features) =
-            mock_external_features(&[CURS, DIST, KERN, ABVM, BLWM, MARK, MKMK]);
-        let markers = make_markers_with_order([]);
-        let mut all = AllLookups::default();
-        let map = all.merge_external_lookups(lookups, &features, &markers);
-
-        assert_eq!(
-            feature_order_for_test(&map, &features),
-            [CURS, KERN, DIST, ABVM, BLWM, MARK, MKMK]
-        );
-    }
-
-    #[test]
-    fn feature_ordering_without_markers_input_order_doesnt_matter_either() {
-        // just to demonstrate that the order we passed in is irrelevant
-        let (lookups, features) =
-            mock_external_features(&[BLWM, KERN, CURS, MKMK, DIST, MARK, ABVM]);
-        let markers = make_markers_with_order([]);
-        let mut all = AllLookups::default();
-        let map = all.merge_external_lookups(lookups, &features, &markers);
-
-        assert_eq!(
-            feature_order_for_test(&map, &features),
-            [CURS, KERN, DIST, ABVM, BLWM, MARK, MKMK]
-        );
-    }
-
-    #[test]
-    fn blwm_with_marker_takes_abvm_with_it() {
-        let (lookups, features) = mock_external_features(&[BLWM, ABVM, DIST]);
-        let markers = make_markers_with_order([BLWM]);
-        let mut all = AllLookups::default();
-        let map = all.merge_external_lookups(lookups, &features, &markers);
-        // because BLWM has a marker and ABVM depends on it, we insert ABVM first
-        assert_eq!(feature_order_for_test(&map, &features), [ABVM, BLWM, DIST]);
-    }
-
-    #[test]
-    fn marks_with_marker_goes_before_kern() {
-        let (lookups, features) = mock_external_features(&[MARK, KERN]);
-        // 'mark' gets an explicit location
-        let markers = make_markers_with_order([MARK]);
-        let mut all = AllLookups::default();
-        let map = all.merge_external_lookups(lookups, &features, &markers);
-        assert_eq!(feature_order_for_test(&map, &features), [MARK, KERN]);
-    }
-
-    #[test]
-    fn mkmk_brings_along_the_whole_family() {
-        let (lookups, features) = mock_external_features(&[BLWM, KERN, MKMK, DIST, MARK, ABVM]);
-        let markers = make_markers_with_order([MKMK]);
-        let mut all = AllLookups::default();
-        let map = all.merge_external_lookups(lookups, &features, &markers);
-        assert_eq!(
-            feature_order_for_test(&map, &features),
-            // abvm/blwm/mark all have to go before mkmk
-            [ABVM, BLWM, MARK, MKMK, KERN, DIST]
-        );
-    }
 }
